Added the rest of the 360 images, Fixed CRUD Styling, Still need to consider about information being able to be deleted such as title/description and terms and conditions. (add padding to mobile crud pages)

+++ app/Http/Controllers/Admin/InfoController.php
@@ -74,7 +74,16 @@ class InfoController extends Controller
 
     public function destroy(Information $info) : RedirectResponse
     {
+        $protectedIds = Information::whereIn('title', ['Title', 'Description'])->pluck('id')->toArray();
+
+        if (in_array($info->id, $protectedIds, true)) {
+            return redirect()
+                ->route('admin.info.index')
+                ->with('info', 'This information entry is protected and cannot be deleted.');
+        }
+
         $info->delete();
+
         return redirect()->route('admin.info.index')->with('success', 'Information deleted.');
     }
 }
